refactor: rename package psqldao to sqldao and fix query config unit test
